+++ src/main/java/org/frc5687/infiniterecharge/robot/Constants.java
@@ -1,8 +1,8 @@
-package frc.robot;
+package org.frc5687.infiniterecharge.robot;
 
 public class Constants {
     public static final int CYCLES_PER_SECOND = 50;
-    public static final int TICKS_PER_UPDATE = 10;
+    public static final int  TICKS_PER_UPDATE = 10;
     public static final double METRIC_FLUSH_PERIOD = 1.0;
 
     public class DriveTrain {
