+++ src/main/java/org/frc5687/infiniterecharge/robot/Robot.java
@@ -1,0 +1,254 @@
+/*----------------------------------------------------------------------------*/
+/* Copyright (c) 2017-2018 FIRST. All Rights Reserved.                        */
+/* Open Source Software - may be modified and shared by FRC teams. The code   */
+/* must be accompanied by the FIRST BSD license file in the root directory of */
+/* the project.                                                               */
+/*----------------------------------------------------------------------------*/
+
+package org.frc5687.infiniterecharge.robot;
+
+import com.kauailabs.navx.frc.AHRS;
+import edu.wpi.first.wpilibj.DriverStation;
+import edu.wpi.first.wpilibj.Notifier;
+import edu.wpi.first.wpilibj.SPI;
+import edu.wpi.first.wpilibj.TimedRobot;
+import edu.wpi.first.wpilibj.controller.PIDController;
+import edu.wpi.first.wpilibj.controller.RamseteController;
+import edu.wpi.first.wpilibj.controller.SimpleMotorFeedforward;
+import edu.wpi.first.wpilibj.geometry.Pose2d;
+import edu.wpi.first.wpilibj.geometry.Rotation2d;
+import edu.wpi.first.wpilibj.geometry.Translation2d;
+import edu.wpi.first.wpilibj.smartdashboard.SmartDashboard;
+import edu.wpi.first.wpilibj.trajectory.Trajectory;
+import edu.wpi.first.wpilibj.trajectory.TrajectoryConfig;
+import edu.wpi.first.wpilibj.trajectory.TrajectoryGenerator;
+import edu.wpi.first.wpilibj.trajectory.constraint.DifferentialDriveVoltageConstraint;
+import edu.wpi.first.wpilibj2.command.Command;
+import edu.wpi.first.wpilibj2.command.CommandScheduler;
+import edu.wpi.first.wpilibj2.command.RamseteCommand;
+import org.frc5687.infiniterecharge.robot.commands.KillAll;
+import org.frc5687.infiniterecharge.robot.subsytems.DriveTrain;
+import org.frc5687.infiniterecharge.robot.subsytems.Shifter;
+import org.frc5687.infiniterecharge.robot.util.*;
+
+import java.io.BufferedReader;
+import java.io.File;
+import java.io.FileReader;
+import java.util.List;
+
+/**
+ * The VM is configured to automatically run this class, and to call the
+ * functions corresponding to each mode, as described in the TimedRobot
+ * documentation. If you change the name of this class or the package after
+ * creating this project, you must also update the build.gradle file in the
+ * project.
+ */
+public class Robot extends OutliersRobot implements ILoggingSource {
+
+    public static IdentityMode identityMode = IdentityMode.competition;
+    private RioLogger.LogLevel _dsLogLevel = RioLogger.LogLevel.warn;
+    private RioLogger.LogLevel _fileLogLevel = RioLogger.LogLevel.warn;
+
+    private int _updateTick = 0;
+
+    private String _name;
+
+    private RobotContainer _robotContainer;
+
+    private boolean _fmsConnected;
+
+    private Command _autoCommand;
+
+    /**
+     * This function is setRollerSpeed when the robot is first started up and should be
+     * used for any initialization code.
+     */
+    @Override
+    public void robotInit() {
+        loadConfigFromUSB();
+        RioLogger.getInstance().init(_fileLogLevel, _dsLogLevel);
+
+
+        metric("Branch", Version.BRANCH);
+        metric("Identity", identityMode.toString());
+        info("Starting " + this.getClass().getCanonicalName() + " from branch " + Version.BRANCH);
+        info("Robot " + _name + " running in " + identityMode.toString() + " mode");
+
+        _robotContainer = new RobotContainer(this);
+        _robotContainer.init();
+
+        // Periodically flushes metrics (might be good to configure enable/disable via USB config file)
+        new Notifier(MetricTracker::flushAll).startPeriodic(Constants.METRIC_FLUSH_PERIOD);
+
+    }
+
+    /**
+     * This function is called every robot packet, no matter the mode. Use
+     * this for items like diagnostics that you want ran during disabled,
+     * autonomous, teleoperated and test.
+     *
+     * <p>This runs after the mode specific periodic functions, but before
+     * LiveWindow and SmartDashboard integrated updating.
+     */
+    @Override
+    public void robotPeriodic() {
+      ourPeriodic();
+    }
+
+    /**
+     * This autonomous (along with the chooser code above) shows how to select
+     * between different autonomous modes using the dashboard. The sendable
+     * chooser code works with the Java SmartDashboard. If you prefer the
+     * LabVIEW Dashboard, remove all of the chooser code and uncomment the
+     * getString line to get the auto name from the text box below the Gyro
+     *
+     * <p>You can add additional auto modes by adding additional comparisons to
+     * the switch structure below with additional strings. If using the
+     * SendableChooser make sure to add them to the chooser code above as well.
+     */
+    @Override
+    public void autonomousInit() {
+        _fmsConnected = DriverStation.getInstance().isFMSAttached();
+        _autoCommand = _robotContainer.getAutonomousCommand();
+
+        if (_autoCommand != null) {
+            _autoCommand.schedule();
+        }
+
+    }
+
+    public void teleopInit() {
+        _fmsConnected = DriverStation.getInstance().isFMSAttached();
+        //_limelight.disableLEDs();
+    }
+
+    /**
+     * This function is called periodically during autonomous.
+     */
+    @Override
+    public void autonomousPeriodic() {
+    }
+
+    /**
+     * This function is called periodically during operator control.
+     */
+    @Override
+    public void teleopPeriodic() {
+    }
+
+    private void ourPeriodic() {
+
+        // Example of starting a new row of metrics for all instrumented objects.
+        // MetricTracker.newMetricRowAll();
+        MetricTracker.newMetricRowAll();
+
+        _robotContainer.periodic();
+
+        CommandScheduler.getInstance().run();
+        update();
+        updateDashboard();
+
+    }
+
+    /**
+     * This function is called periodically during test mode.
+     */
+    @Override
+    public void testPeriodic() {
+        CommandScheduler.getInstance().run();
+    }
+
+    @Override
+    public void disabledInit() {
+        //_limelight.disableLEDs();
+        RioLogger.getInstance().forceSync();
+        RioLogger.getInstance().close();
+//        MetricTracker.flushAll();
+    }
+
+
+    public void updateDashboard() {
+        _updateTick++;
+        if (_updateTick >= Constants.TICKS_PER_UPDATE) {
+            _updateTick = 0;
+            _robotContainer.updateDashboard();
+        }
+    }
+
+
+    private void loadConfigFromUSB() {
+        String output_dir = "/U/"; // USB drive is mounted to /U on roboRIO
+        try {
+            String usbDir = "/U/"; // USB drive is mounted to /U on roboRIO
+            String configFileName = usbDir + "frc5687.cfg";
+            File configFile = new File(configFileName);
+            FileReader reader = new FileReader(configFile);
+            BufferedReader bufferedReader = new BufferedReader(reader);
+
+            String line;
+            while ((line = bufferedReader.readLine()) != null) {
+                processConfigLine(line);
+            }
+
+            bufferedReader.close();
+            reader.close();
+        } catch (Exception e) {
+            identityMode = IdentityMode.competition;
+        }
+    }
+
+    private void processConfigLine(String line) {
+        try {
+            if (line.startsWith("#")) {
+                return;
+            }
+            String[] a = line.split("=");
+            if (a.length == 2) {
+                String key = a[0].trim().toLowerCase();
+                String value = a[1].trim();
+                switch (key) {
+                    case "name":
+                        _name = value;
+                        break;
+                    case "mode":
+                        identityMode = IdentityMode.valueOf(value.toLowerCase());
+                        break;
+                    case "fileloglevel":
+                        _fileLogLevel = RioLogger.LogLevel.valueOf(value.toLowerCase());
+                        break;
+                    case "dsloglevel":
+                        _dsLogLevel = RioLogger.LogLevel.valueOf(value.toLowerCase());
+                        break;
+                }
+            }
+        } catch (Exception e) {
+
+        }
+    }
+
+    private void update() {
+
+    }
+
+    public enum IdentityMode {
+        competition(0),
+        practice(1),
+        programming(2);
+
+        private int _value;
+
+        IdentityMode(int value) {
+            this._value = value;
+        }
+
+        public int getValue() {
+            return _value;
+        }
+    }
+
+    public IdentityMode getIdentityMode() {
+        return identityMode;
+    }
+
+
+}
