+++ src/main/java/org/frc5687/infiniterecharge/robot/RobotMap.java
@@ -1,4 +1,4 @@
-package frc.robot;
+package org.frc5687.infiniterecharge.robot;
 
 public class RobotMap {
