+++ src/main/java/org/frc5687/infiniterecharge/robot/subsytems/OutliersSubsystem.java
@@ -1,20 +1,17 @@
-package frc.robot.commands;
+package org.frc5687.infiniterecharge.robot.subsytems;
 
-import edu.wpi.first.wpilibj2.command.Command;
 import edu.wpi.first.wpilibj.smartdashboard.SmartDashboard;
-import edu.wpi.first.wpilibj2.command.CommandBase;
-import frc.robot.util.ILoggingSource;
-import frc.robot.util.MetricTracker;
-import frc.robot.util.RioLogger;
+import edu.wpi.first.wpilibj2.command.SubsystemBase;
+import org.frc5687.infiniterecharge.robot.util.ILoggingSource;
+import org.frc5687.infiniterecharge.robot.util.MetricTracker;
+import org.frc5687.infiniterecharge.robot.util.OutliersContainer;
+import org.frc5687.infiniterecharge.robot.util.RioLogger;
 
-public abstract class OutliersCommand extends CommandBase implements ILoggingSource {
+public abstract class OutliersSubsystem extends SubsystemBase implements ILoggingSource {
     private MetricTracker _metricTracker;
 
-    public OutliersCommand() {
-    }
-
-    public OutliersCommand(double timeout) {
-        super.withTimeout(timeout);
+    public OutliersSubsystem(OutliersContainer container) {
+        container.registerSubSystem(this);
     }
 
     @Override
@@ -39,58 +36,39 @@ public abstract class OutliersCommand extends CommandBase implements ILoggingSou
 
     public void metric(String name, String value) {
         SmartDashboard.putString(getClass().getSimpleName() + "/" + name, value);
-        if (_metricTracker != null) {
+        if (_metricTracker!=null) {
             _metricTracker.put(name, value);
         }
     }
 
     public void metric(String name, double value) {
         SmartDashboard.putNumber(getClass().getSimpleName() + "/" + name, value);
-        if (_metricTracker != null) {
+        if (_metricTracker!=null) {
             _metricTracker.put(name, value);
         }
     }
 
     public void metric(String name, boolean value) {
         SmartDashboard.putBoolean(getClass().getSimpleName() + "/" + name, value);
-        if (_metricTracker != null) {
+        if (_metricTracker!=null) {
             _metricTracker.put(name, value);
         }
     }
 
+    // Example of metrics collection. In a child class's constructor:
+    // getMetricTracker().registerReportableMetricName("foo");
+    // getMetricTracker().registerReportableMetricName("bar");
+    // getMetricTracker().registerReportableMetricName("baz");
+    //
+    // Later on in the child class...
+    // metric("foo", 123);
+    // metric("bar", "elvis");
+    // metric("baz", 42.42);
+    // metric("pants", 99);    <~ This metric won't get written to USB storage because it wasn't registered.
+
     protected void logMetrics(String... metrics) {
         _metricTracker = MetricTracker.createMetricTracker(getClass().getSimpleName(), metrics);
-        _metricTracker.pause();
     }
 
-    @Override
-    public void initialize() {
-        super.initialize();
-        if (_metricTracker != null) {
-            _metricTracker.resume();
-        }
-    }
-
-    @Override
-    public void end(boolean interrupted) {
-        super.end(interrupted);
-        if (_metricTracker != null) {
-            _metricTracker.pause();
-        }
-    }
-
-
-    private long _start;
-
-    @Override
-    public void execute() {
-        if (_metricTracker != null && _metricTracker.isPaused()) {
-            _metricTracker.resume();
-        }
-
-    }
-
-    protected void innerExecute() {
-
-    }
+    public abstract void updateDashboard();
 }
