+++ src/main/java/org/frc5687/infiniterecharge/robot/util/Gamepad.java
@@ -1,4 +1,4 @@
-package frc.robot.util;
+package org.frc5687.infiniterecharge.robot.util;
 
 import edu.wpi.first.wpilibj.Joystick;
