+++ src/main/java/org/frc5687/infiniterecharge/robot/util/OutliersRobot.java
@@ -1,0 +1,41 @@
+package org.frc5687.infiniterecharge.robot.util;
+
+import edu.wpi.first.wpilibj.TimedRobot;
+import edu.wpi.first.wpilibj.smartdashboard.SmartDashboard;
+
+public abstract class OutliersRobot extends TimedRobot implements ILoggingSource {
+
+
+    public void metric(String name, boolean value) {
+        SmartDashboard.putBoolean(getClass().getSimpleName() + "/" + name, value);
+    }
+
+    public void metric(String name, String value) {
+        SmartDashboard.putString(getClass().getSimpleName() + "/" + name, value);
+    }
+
+    public void metric(String name, double value) {
+        SmartDashboard.putNumber(getClass().getSimpleName() + "/" + name, value);
+    }
+
+    @Override
+    public void error(String message) {
+        RioLogger.error(this, message);
+    }
+
+    @Override
+    public void warn(String message) {
+        RioLogger.warn(this, message);
+    }
+
+    @Override
+    public void info(String message) {
+        RioLogger.info(this, message);
+    }
+
+    @Override
+    public void debug(String message) {
+        RioLogger.debug(this, message);
+    }
+
+}
