+++ src/main/java/org/frc5687/infiniterecharge/robot/util/PDP.java
@@ -1,4 +1,4 @@
-package frc.robot.util;
+package org.frc5687.infiniterecharge.robot.util;
 
 import edu.wpi.first.wpilibj.PowerDistributionPanel;
 import edu.wpi.first.wpilibj.smartdashboard.SmartDashboard;
