+++ src/main/java/org/frc5687/infiniterecharge/robot/util/Version.java
@@ -1,4 +1,4 @@
-package frc.robot.util;
+package org.frc5687.infiniterecharge.robot.util;
 
 /**
  * A version class to read current Git information from deployed robot code
@@ -6,7 +6,7 @@ package frc.robot.util;
  */
 
 public class Version {
-    public static final String BRANCH = "Shooter/#5-CreateSubsystem";
-    public static final String REVISION = "912611f";
+    public static final String BRANCH = "DriveTrain/#2-CreateSubsystems";
+    public static final String REVISION = "5da2009";
     public static final String MODIFIED = " *";
 }
